Add method DataRegionSelection.getSelectedIntegers() and eliminate lots of duplicate String -> Integer parsing code
TextWriter now implements AutoCloseable -> switch callers to use try-with-resources

Switch to try-with-resources elsewhere, eliminate unused imports and other warnings

+++ study/src/org/labkey/study/controllers/StudyController.java
@@ -3617,7 +3617,7 @@ public class StudyController extends BaseStudyController
             QCState newState = null;
             if (updateQCForm.getNewState() != null)
             {
-                newState = StudyManager.getInstance().getQCStateForRowId(getContainer(), updateQCForm.getNewState().intValue());
+                newState = StudyManager.getInstance().getQCStateForRowId(getContainer(), updateQCForm.getNewState());
                 if (newState == null)
                 {
                     errors.reject(null, "The selected state could not be found.  It may have been deleted from the database.");
